app/api/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.30.6
asyncpg==0.29.0
httpx==0.27.2
python-dotenv==1.0.1
prometheus-fastapi-instrumentator==7.0.0
prometheus-client==0.20.0